Fix same algorithm reference across different instances and repeatable random seeds
When an algorithm instance is passed to FrameworkConf this is creating N references to the same object
which means that we run exactly the same algorithm across all the instances, with the same seed as well.
Introducing a factory pattern to create different objects as well as seeding the random generating in them
independently fix the issue and allow repeatable different executions across different cores.
See LocalApp example for usage.